Add authentication before all actions except login and account creation
