+++ app/controllers/application_controller.rb
@@ -1,2 +1,5 @@
 class ApplicationController < ActionController::Base
+  include Authentication
+
+  before_action :authenticate_staff_user
 end
