+++ app/controllers/sessions_controller.rb
@@ -1,4 +1,6 @@
 class SessionsController < ApplicationController
+  skip_before_action :authenticate_staff_user, only: [:create]
+
   def create
     username = params[:username]
     password = params[:password]
